requirements.txt
redis>=4.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
flask>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
